requirements.txt
langchain
langchain-community
langchain-ollama
langgraph
chromadb
pymupdf
streamlit
sentence-transformers
python-dotenv
tiktoken
